Refactored method_body_typechecker and extracted expr_typechecker.rs.
